reccomendationEngine/pyproject.toml
[project]
name = "embedderserver"
version = "0.1.0"
description = "Inngest-powered audio embedding service for OneMelody"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.30.0",
    "inngest>=0.5.17",
    "essentia>=2.1b6.dev1110",
    "recombee-api-client>=5.0.0",
    "boto3>=1.34.0",
    "numpy>=1.26.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.1",
]